Homework 2-1 LinkedDeque
(Deque) A double-ended queue or deque (pronounced “deck”) is a generalization of a stack and a queue that supports adding and removing items from either the front or the back of the data structure. Creates a generic iterable data type LinkedDeque<Item> in LinkedDeque.java that uses a linked list to implement.