Homework 3-3 Autocomplete
Implements a data type that provides autocomplete functionality for a given set of string and weights, using Term and BinarySearchDeluxe. Sorts the terms in lexicographic order; uses binary search to ﬁnd the set of terms that start with a given preﬁx; and sorts the matching terms in descending order by weight.