Create Homework 5-1
The purpose of this assignment is to create a symbol table data type whose keys are two-dimensional points. We used a 2d-tree to support eﬃcient range search (ﬁnd all the points contained in a query rectangle) and k-nearest neighbor search (ﬁnd k points that are closest to a query point).