Homework 6-1 WordNet
WordNet is a semantic lexicon for the English language that computational linguists and cognitive scientists use extensively. For example, WordNet was a key component in IBM’s Jeopardy-playing Watson computer system. WordNet groups words into sets of synonyms called synsets. For example, {AND circuit, AND gate} is a synset that represent a logical gate that ﬁres only when all of its inputs ﬁre. WordNet also describes semantic relationships between synsets. One such relationship is the is-a relationship, which connects a hyponym (more speciﬁc synset) to a hypernym (more general synset). For example, the synset {gate, logic gate} is a hypernym of {AND circuit, AND gate} because an AND gate is a kind of logic gate.